day 24, "bruteforce" way to find the velocity, I know there is a smarter way, but iterating over 2k elements is not that bad. refactor, move all tests and code for part 2 to separate class

+++ utility/src/main/kotlin/com/github/shmvanhouten/adventofcode/utility/compositenumber/CompositeNumbers.kt
@@ -1,6 +1,5 @@
 package com.github.shmvanhouten.adventofcode.utility.compositenumber
 
-import java.math.BigDecimal
 import java.math.BigInteger
 
 fun primeFactors(number: Long): List<Long> {
@@ -18,13 +17,13 @@ fun primeFactors(number: Long): List<Long> {
     return factors
 }
 
-fun primeFactors(number: BigDecimal): List<BigDecimal> {
-    val factors = mutableListOf<BigDecimal>()
+fun primeFactors(number: BigInteger): List<BigInteger> {
+    val factors = mutableListOf<BigInteger>()
     var remainder = number
 
-    var divisor = 2.toBigDecimal()
-    while (remainder > BigDecimal.ONE) {
-        while (remainder % divisor == BigDecimal.ZERO) {
+    var divisor = 2.toBigInteger()
+    while (remainder > BigInteger.ONE) {
+        while (remainder % divisor == BigInteger.ZERO) {
             factors.add(divisor)
             remainder /= divisor
         }
